Change subTrackId from byte to short
* Release 10.526.741 carries 1.300 subTracks

+++ de.tgmz.discogs/src/test/java/de/tgmz/discogs/test/DiscogsTest.java
@@ -211,8 +211,8 @@ public class DiscogsTest {
 		sti0.setTrack(t0);
 		sti1.setTrack(t1);
 
-		sti0.setSubTrackNumber((byte) 3);
-		sti1.setSubTrackNumber((byte) 3);
+		sti0.setSubTrackNumber((short) 3);
+		sti1.setSubTrackNumber((short) 3);
 		
 		// Ensure that subTracks are equal iff the tracks _ids_ and subtracknumbers are equal
 		assertEquals(sti0, sti1);
